Candidate backend done bugs may occur

+++ ApplicationLayer/Interfaces/IUser.cs
@@ -13,10 +13,10 @@ namespace ApplicationLayer.Interfaces
         Task<ResponseVM> Login(LoginUserRequest user);
         Task<ResponseVM> AddRecuriter(AddUserRequest user);
         Task<ResponseVM> GetAllUsers();
-        Task<ResponseVM> GetUserById(int id);
+        Task<ResponseVM> GetUserById(long id);
         Task<ResponseVM> UpdateUser(UpdateUserRequest request);
         Task<ResponseVM> PatchUserAsync(PatchUserRequest request);
-        Task<ResponseVM> DeleteUser(int id);
+        Task<ResponseVM> DeleteUser(long id);
         Task<ResponseVM> UpdateUserStatusAsync(UpdateUserStatusRequest request);
     }
 }
